+++ Engine/Source/Engine/Graphics/Primitives/Primitive.h
@@ -468,7 +468,7 @@ namespace Engine::Primitive
 					}
 				}
 
-				int n = vertices->size();
+				int n = (int)vertices->size();
 
 				const float latitude = ((float)effectiveJ * PI) / (float)meridians - PI / 2;
 
@@ -541,7 +541,7 @@ namespace Engine::Primitive
 
 		for (int i = 0; i < 2; ++i)
 		{
-			float h = -1.0 / 2.0f + i * 1.0; // z value; -h/2 to h/2
+			float h = -1.0f / 2.0f + (float)i * 1.0f; // z value; -h/2 to h/2
 			float t = 1.0f - i; // vertical tex coord; 1 to 0
 
 			for (int j = 0, k = 0; j <= sectors; ++j, k += 3)
